requirements.txt
qiskit-terra>=0.21.2
qiskit-aer>=0.11.0
pybind11<=2.9.1
PyMatching>=0.6.0,!=2.0.0
rustworkx>=0.12.0
networkx>=2.6.3
sympy>=1.9
numpy>=1.21.0
ipython
ipywidgets==7.7.2
